+++ api-types/src/lib.rs
@@ -29,7 +29,7 @@ pub use vm_config::cpus_config::{
     CpusConfigParseError, ParseCoreSchedulingError,
 };
 pub use vm_config::disk_config::VirtQueueAffinity;
-pub use vm_config::memory_config::HotplugMethod;
+pub use vm_config::memory_config::{HotplugMethod, MemoryZoneConfig};
 pub use vm_config::net_config::{ParseVhostModeError, VhostMode};
 pub use vm_config::numa_config::NumaDistance;
 pub use vm_coredump::VmCoredumpData;
