+++ vmm/src/lib.rs
@@ -3429,7 +3429,7 @@ mod unit_tests {
 
     use api_types::{
         ConsoleOutputMode, CoreScheduling, CpuFeatures, CpusConfig, HotplugMethod,
-        RestoredVfioConfig,
+        MemoryZoneConfig, RestoredVfioConfig,
     };
     use arch::CpuProfile;
 
@@ -3437,8 +3437,8 @@ mod unit_tests {
     #[cfg(target_arch = "x86_64")]
     use crate::vm_config::DebugConsoleConfig;
     use crate::vm_config::{
-        CommonConsoleConfig, ConsoleConfig, MemoryConfig, MemoryZoneConfig, PayloadConfig,
-        PciDeviceCommonConfig, PlatformConfig, RngConfig, SerialConfig,
+        CommonConsoleConfig, ConsoleConfig, MemoryConfig, PayloadConfig, PciDeviceCommonConfig,
+        PlatformConfig, RngConfig, SerialConfig,
     };
 
     fn create_dummy_vmm() -> Vmm {
